Multitenant/requirements.txt
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
Babel==2.14.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
cachetools==5.3.3
catfish==4.16.4
certifi==2024.2.2
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
configobj==5.0.8
configparser==7.0.0
constantly==23.10.4
contourpy==1.2.1
crispy-bootstrap5==2024.2
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.2
dbus-python==1.3.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.4
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-otp==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
etelemetry==0.3.1
executing==2.0.1
ffmpeg-python==0.2.0
filelock==3.14.0
fitz==0.0.1.dev2
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.53.0
frontend==0.0.3
future==1.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
gyp-next==0.16.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpie==3.2.2
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
ipykernel==6.29.2
ipython==8.21.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kali-tweaks==2023.3.2
kiwisolver==1.4.5
lightdm-gtk-greeter-settings==1.2.3
looseversion==1.3.0
louis==3.30.0
lxml==5.2.2
Mako==1.3.2
mariadb==1.1.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mechanize==0.4.10
multidict==6.0.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
onboard==1.4.1
openai==1.30.1
packaging==23.2
pandas==2.2.1
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.0
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.26.0
pydantic==2.7.1
pydantic_core==2.18.2
pydot==2.0.0
Pygments==2.17.2
PyGObject==3.47.0.dev0
pyinotify==0.9.6
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
python-apt==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.28
pyxnat==1.6.2
pyzmq==25.1.2
rdflib==7.0.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
scipy==1.13.1
seaborn==0.13.2
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
traits==6.3.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
ufw==0.36.2
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
xdg==5
zope.interface==7.0.2
